requirements.txt
python-telegram-bot[ext]==21.6
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
pytz==2024.1
asyncpg==0.29.0
openpyxl==3.1.2
aiohttp==3.9.5
aiohttp-cors==0.7.0
openai==1.35.13
httpx==0.27.2
oauth2client==4.1.3
pandas==2.2.2
